Improved async task result retrieval.
controller.get_result now accepts a request-id (as returned
by worker.request) and does NOT implement blocking semantics.
It's up to the user to wait-loop or whatever if they want to
synchronized on a task's completion.

Internally, a managed (cf multiprocessing mod) dict is used to
store the reqid and the completed task